[Fix] resolved pr issues after review
