+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { UserWarning } from './UserWarning';
 import { USER_ID } from './api/todos';
-import { TodoApp } from './components/todoApp';
+import { TodoApp } from './components/TodoApp';
 
 export const App: React.FC = () => {
   if (!USER_ID) {
