+++ src/components/Notification/Notification.tsx
@@ -1,0 +1,36 @@
+import cn from 'classnames';
+
+interface Props {
+  error: string | null;
+  clearError: VoidFunction;
+}
+
+export const Notification: React.FC<Props> = ({ error, clearError }) => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className={cn('notification is-danger is-light has-text-weight-normal', {
+        hidden: !error,
+      })}
+    >
+      <button
+        data-cy="HideErrorButton"
+        type="button"
+        className="delete"
+        onClick={clearError}
+      />
+
+      {error}
+      {/* show only one message at a time */}
+      {/* Unable to load todos
+        <br />
+        Title should not be empty
+        <br />
+        Unable to add a todo
+        <br />
+        Unable to delete a todo
+        <br />
+        Unable to update a todo */}
+    </div>
+  );
+};
