+++ src/components/Notification/index.ts
@@ -1,0 +1,1 @@
+export { Notification } from './Notification';
